+++ lib/capybara/poltergeist/client.rb
@@ -57,32 +57,16 @@ module Capybara::Poltergeist
 
       redirect_stdout do
         @pid = Process.spawn(*command.map(&:to_s), process_options)
-        ObjectSpace.define_finalizer(self, self.class.process_killer(@pid) )
+        ObjectSpace.define_finalizer(self, self.class.process_killer(@pid))
       end
     end
 
     def stop
       if pid
-        begin
-          if Capybara::Poltergeist.windows?
-            Process.kill('KILL', pid)
-          else
-            Process.kill('TERM', pid)
-            begin
-              Timeout.timeout(KILL_TIMEOUT) { Process.wait(pid) }
-            rescue Timeout::Error
-              Process.kill('KILL', pid)
-              Process.wait(pid)
-            end
-          end
-        rescue Errno::ESRCH, Errno::ECHILD
-          # Zed's dead, baby
-        end
+        kill_phantomjs
         @out_thread.kill
-        @write_io.close
-        @read_io.close
+        close_io
         ObjectSpace.undefine_finalizer(self)
-        @pid = nil
       end
     end
 
@@ -103,7 +87,10 @@ module Capybara::Poltergeist
     private
 
     # This abomination is because JRuby doesn't support the :out option of
-    # Process.spawn
+    # Process.spawn. To be honest it works pretty bad with pipes too, because
+    # we ought close writing end in parent process immediately but JRuby will
+    # lose all the output from child. Process.popen can be used here and seems
+    # it works with JRuby but I've experienced strange mistakes on Rubinius.
     def redirect_stdout
       prev = STDOUT.dup
       prev.autoclose = false
@@ -113,6 +100,46 @@ module Capybara::Poltergeist
     ensure
       STDOUT.reopen(prev)
       $stdout = STDOUT
+    end
+
+    def kill_phantomjs
+      begin
+        if Capybara::Poltergeist.windows?
+          Process.kill('KILL', pid)
+        else
+          Process.kill('TERM', pid)
+          begin
+            Timeout.timeout(KILL_TIMEOUT) { Process.wait(pid) }
+          rescue Timeout::Error
+            Process.kill('KILL', pid)
+            Process.wait(pid)
+          end
+        end
+      rescue Errno::ESRCH, Errno::ECHILD
+        # Zed's dead, baby
+      end
+      @pid = nil
+    end
+
+    # We grab all the output from PhantomJS like console.log in another thread
+    # and when PhantomJS crashes we try to restart it. In order to do it we stop
+    # server and client and on JRuby see this error `IOError: Stream closed`.
+    # It happens because JRuby tries to close pipe and it is blocked on `eof?`
+    # or `readpartial` call. The error is raised in the related thread and it's
+    # not actually main thread but the thread that listens to the output. That's
+    # why if you put some debug code after `rescue IOError` it won't be shown.
+    # In fact the main thread will continue working after the error even if we
+    # don't use `rescue`. The first attempt to fix it was a try not to block on
+    # IO, but looks like similar issue appers after JRuby upgrade. Perhaps the
+    # only way to fix it is catching the exception what this method overall does.
+    def close_io
+      [@write_io, @read_io].each do |io|
+        begin
+          io.close unless io.closed?
+        rescue IOError
+          raise unless RUBY_ENGINE == 'jruby'
+        end
+      end
     end
   end
 end